misc(sink): Refactor + remove dead code
