--- sink/src/kg/client.rs
@@ -1,154 +1,0 @@
-use crate::bootstrap::constants;
-
-use sdk::{
-    error::DatabaseError,
-    mapping::{Entity, Relation},
-    models::{self, BlockMetadata},
-    pb,
-};
-
-#[derive(Clone)]
-pub struct Client {
-    pub neo4j: neo4rs::Graph,
-}
-
-impl Client {
-    pub async fn new(uri: &str, user: &str, pass: &str) -> anyhow::Result<Self> {
-        let neo4j = neo4rs::Graph::new(uri, user, pass).await?;
-        Ok(Self { neo4j })
-    }
-
-    /// Bootstrap the database with the initial data
-    // pub async fn bootstrap(&self, _rollup: bool) -> Result<(), DatabaseError> {
-    //     models::Space::builder(
-    //         constants::ROOT_SPACE_ID,
-    //         constants::ROOT_SPACE_DAO_ADDRESS,
-    //         &BlockMetadata::default(),
-    //     )
-    //     .space_plugin_address(constants::ROOT_SPACE_PLUGIN_ADDRESS)
-    //     .voting_plugin_address(constants::ROOT_SPACE_MAIN_VOTING_ADDRESS)
-    //     .member_access_plugin(constants::ROOT_SPACE_MEMBER_ACCESS_ADDRESS)
-    //     .build()
-    //     .upsert(&self.neo4j)
-    //     .await
-
-    //     // self.process_ops(
-    //     //     &BlockMetadata::default(),
-    //     //     constants::ROOT_SPACE_ID,
-    //     //     bootstrap::bootstrap(),
-    //     // )
-    //     // .await
-    // }
-
-    /// Reset the database by deleting all nodes and relations and re-bootstrapping it
-    pub async fn reset_db(&self, rollup: bool) -> anyhow::Result<()> {
-        // Delete all nodes and relations
-        let mut txn = self.neo4j.start_txn().await?;
-        txn.run(neo4rs::query("MATCH (n) DETACH DELETE n")).await?;
-        txn.commit().await?;
-
-        // Re-bootstrap the database
-        // self.bootstrap(rollup).await?;
-
-        Ok(())
-    }
-
-    pub async fn process_ops(
-        &self,
-        block: &models::BlockMetadata,
-        space_id: &str,
-        ops: impl IntoIterator<Item = pb::ipfs::Op>,
-    ) -> Result<(), DatabaseError> {
-        for op in ops {
-            match (op.r#type(), op) {
-                (
-                    pb::ipfs::OpType::SetTriple,
-                    pb::ipfs::Op {
-                        triple:
-                            Some(pb::ipfs::Triple {
-                                entity,
-                                attribute,
-                                value: Some(value),
-                            }),
-                        ..
-                    },
-                ) => {
-                    tracing::info!("SetTriple: {}, {}, {:?}", entity, attribute, value,);
-
-                    Entity::<()>::set_triple(
-                        &self.neo4j,
-                        block,
-                        space_id,
-                        &entity,
-                        &attribute,
-                        &value,
-                    )
-                    .await?
-                }
-                (
-                    pb::ipfs::OpType::DeleteTriple,
-                    pb::ipfs::Op {
-                        triple: Some(triple),
-                        ..
-                    },
-                ) => {
-                    tracing::info!(
-                        "DeleteTriple: {}, {}, {:?}",
-                        triple.entity,
-                        triple.attribute,
-                        triple.value,
-                    );
-
-                    Entity::<()>::delete_triple(&self.neo4j, block, space_id, triple).await?
-                }
-                // TODO: Handle these cases
-                // (pb::ipfs::OpType::SetTripleBatch, op) => {
-                // }
-                // (pb::ipfs::OpType::DeleteEntity, op) => {
-                // }
-                (
-                    pb::ipfs::OpType::CreateRelation,
-                    pb::ipfs::Op {
-                        relation: Some(relation),
-                        ..
-                    },
-                ) => {
-                    tracing::info!(
-                        "CreateRelation: {}, {}, {}, {}",
-                        relation.id,
-                        relation.r#type,
-                        relation.from_entity,
-                        relation.to_entity,
-                    );
-
-                    Relation::<()>::new(
-                        &relation.id,
-                        space_id,
-                        &relation.r#type,
-                        &relation.from_entity,
-                        &relation.to_entity,
-                        block,
-                        (),
-                    )
-                    .upsert(&self.neo4j)
-                    .await?
-                }
-                (
-                    pb::ipfs::OpType::DeleteRelation,
-                    pb::ipfs::Op {
-                        relation: Some(relation),
-                        ..
-                    },
-                ) => {
-                    tracing::info!("DeleteRelation: {}", relation.id);
-                    Entity::<()>::delete(&self.neo4j, block, &relation.id, space_id).await?
-                }
-                (typ, maybe_triple) => {
-                    tracing::warn!("Unhandled case: {:?} {:?}", typ, maybe_triple);
-                }
-            }
-        }
-
-        Ok(())
-    }
-}
