--- sink/src/kg/mod.rs
@@ -1,3 +1,0 @@
-pub mod client;
-
-pub use client::Client;
