+++ sink/src/lib.rs
@@ -1,4 +1,3 @@
 pub mod bootstrap;
 pub mod events;
-pub mod kg;
 pub mod metrics;
