+++ sink/src/main.rs
@@ -3,10 +3,9 @@ use std::env;
 use anyhow::Error;
 use axum::{response::Json, routing::get, Router};
 use clap::{Args, Parser};
-use sink::{events::EventHandler, kg, metrics};
+use sink::{events::EventHandler, metrics};
 use substreams_utils::Sink;
 use tracing::Level;
-use tracing_subscriber::fmt::format::FmtSpan;
 use tracing_subscriber::layer::SubscriberExt;
 use tracing_subscriber::util::SubscriberInitExt;
 
@@ -41,7 +40,7 @@ async fn main() -> Result<(), Error> {
     set_log_level();
     init_tracing(args.log_file);
 
-    let kg_client = kg::Client::new(
+    let neo4j = neo4rs::Graph::new(
         &args.neo4j_args.neo4j_uri,
         &args.neo4j_args.neo4j_user,
         &args.neo4j_args.neo4j_pass,
@@ -49,10 +48,10 @@ async fn main() -> Result<(), Error> {
     .await?;
 
     if args.reset_db {
-        kg_client.reset_db(args.rollup).await?;
+        reset_db(&neo4j).await?;
     };
 
-    let sink = EventHandler::new(kg_client);
+    let sink = EventHandler::new(neo4j);
 
     start_http_server().await;
 
@@ -106,13 +105,25 @@ struct Neo4jArgs {
     neo4j_pass: String,
 }
 
+pub async fn reset_db(neo4j: &neo4rs::Graph) -> anyhow::Result<()> {
+    // Delete all nodes and relations
+    let mut txn = neo4j.start_txn().await?;
+    txn.run(neo4rs::query("MATCH (n) DETACH DELETE n")).await?;
+    txn.commit().await?;
+
+    Ok(())
+}
+
 fn init_tracing(log_file: Option<String>) {
     if let Some(log_file) = log_file {
         // Set the path of the log file
         let now = chrono::Utc::now();
-        let file_appender = tracing_appender::rolling::never(".", format!("{}-{log_file}", now.format("%Y-%m-%d-%H-%M-%S")),);
+        let file_appender = tracing_appender::rolling::never(
+            ".",
+            format!("{}-{log_file}", now.format("%Y-%m-%d-%H-%M-%S")),
+        );
         let (non_blocking, _guard) = tracing_appender::non_blocking(file_appender);
-        
+
         tracing_subscriber::fmt::fmt()
             .with_max_level(Level::INFO)
             .with_writer(non_blocking)
@@ -126,7 +137,6 @@ fn init_tracing(log_file: Option<String>) {
             .with(tracing_subscriber::fmt::layer())
             .init();
     }
-
 }
 
 fn set_log_level() {
